requirements.txt
colorama
python-dotenv
pathlib
tqdm
send2trash
pytest
pytest-cov